+++ globals.py
@@ -1,0 +1,1 @@
+HTTP_TIMEOUT = 30
